【宗门】宗门招募弟子逻辑

+++ HEDAO/Assets/Scripts/Save/SaveData.cs
@@ -6,10 +6,10 @@ using UnityEngine;
 public class SaveData
 {
     public string SaveName = DateTime.Now.ToShortDateString();
-    public List<int> RoleList = new List<int>();
+    public List<long> RoleList = new List<long>();
     public List<Role> RandomRoleList = null;
-    public Dictionary<int, Role> DiscipleList = new Dictionary<int, Role>();
-    public HashSet<int> RoleTeamSet = new HashSet<int>();
+    public Dictionary<long, Role> DiscipleList = new Dictionary<long, Role>();
+    public HashSet<long> RoleTeamSet = new HashSet<long>();
 
     public SaveData()
     {
